ls.rbの処理の内容をクラス化
