+++ 07.ls_object/lib/entry_list.rb
@@ -1,5 +1,7 @@
 # frozen_string_literal: true
 
+require_relative 'entry'
+
 class EntryList
   attr_reader :entries
